requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.1
python-multipart>=0.0.6
jinja2>=3.1.2
websockets>=11.0.3
pydantic>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.19
PyJWT>=2.8.0
